requirements.txt
functions-framework==3.*
Flask==2.2.2
Werkzeug==2.3.7
Pillow==9.5.0
numpy==1.26.4
opencv-python-headless
requests
python-dotenv
easyocr==1.7.1
ultralytics==8.2.18
gcsfs
google-cloud-storage